pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brain"
version = "0.1.0"
description = "Markdown vault CLI + HTTP server for JarvisBrain"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "httpx>=0.27",
]

[project.scripts]
brain = "brain.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["brain"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.coverage.run]
source = ["brain"]
branch = true
